Add delay in `_write_byte` method
This commit introduces a small delay of 1 millisecond in the
`_write_byte` method of the `PCF8574` class. The method sends an I2C
command using MicroPython to a PCF8574 chip to control an LCD using the
HD44780 in 4-bit mode.

The `_write_byte` method was previously writing to the I2C bus without
checking the Busy Flag (BF) as indicated in the HD44780U datasheet.
According to the datasheet, the BF must be checked after the 4-bit data
has been transferred twice, to ensure that the previous operation has
completed before the next instruction is accepted.

However, due to the absence of this check in the implementation, there
may be instances of unexpected behavior, like the LCD "freezing" during
operation. This could be due to writing to the I2C bus too quickly,
before the previous operation has finished.

As a quick patch to mitigate this issue, a delay of 1 millisecond is
introduced using the `utime.sleep_ms(1)` function call before writing to
the I2C bus.

It's important to note that this is a temporary fix and does not address
the issue when the Raspberry Pi Pico is running without communicating to
a host machine via UART over USB. The UART communication may provide an
artificial delay that compensates for not checking the Busy Flag. The
ideal solution would be to adhere to the datasheet's recommendation and
implement the Busy Flag check.